subject_analyzer/requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv==1.0.0
rich==13.7.0
tavily-python==0.2.6 
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.0.0
tavily-python==0.2.6
requests==2.31.0
beautifulsoup4==4.12.2
rich==13.7.0
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1 